+++ src/egeo-demo/st-sidebar-demo/st-sidebar-demo.component.ts
@@ -19,10 +19,26 @@ import { StSidebarItem } from '@stratio/egeo';
 
 export class StSidebarDemoComponent {
    public items: StSidebarItem[] = [
-      {id: 'vault-roles', label: 'Vault Roles'},
-      {id: 'identities', label: 'Identities', class: 'warning'},
-      {id: 'masters', label: 'Masters'},
-      {id: 'agents', label: 'Agents'},
-      {id: 'roles', label: 'Roles'}
+      { id: 'vault-roles', label: 'Vault Roles' },
+      { id: 'identities', label: 'Identities', class: 'warning' },
+      {
+         id: 'masters',
+         label: 'Masters',
+         items: [{
+            id: 'sub-item1',
+            label: 'Subitem 1',
+            items: [
+               { id: 'sub-item1.1', label: 'Subitem 1.1' },
+               { id: 'sub-item1.2', label: 'Subitem 1.2' }]
+         }, { id: 'sub-item2', label: 'Subitem 2' }]
+      },
+      { id: 'agents', label: 'Agents' },
+      { id: 'roles', label: 'Roles' }
    ];
+
+   public activeItem: string;
+
+   onChangeActive(item: string) {
+      this.activeItem = item;
+   }
 }