+++ src/lib/st-sidebar/st-sidebar-item.interface.ts
@@ -13,4 +13,5 @@ export interface StSidebarItem {
    id: string;
    label: string;
    class?: string;
+   items?: StSidebarItem[];
 }