+++ src/lib/st-sidebar/st-sidebar.component.spec.ts
@@ -11,6 +11,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ChangeDetectionStrategy } from '@angular/core';
 import { StSidebarComponent } from './st-sidebar.component';
+import { StSidebarItemListComponent } from './st-sidebar-item-list/st-sidebar-item-list.component';
 
 describe('StSidebar', () => {
 
@@ -21,10 +22,13 @@ describe('StSidebar', () => {
    beforeEach(async(() => {
       TestBed.configureTestingModule({
          imports: [],
-         declarations: [StSidebarComponent]
+         declarations: [StSidebarComponent, StSidebarItemListComponent]
       })
       // remove this block when the issue #12313 of Angular is fixed
          .overrideComponent(StSidebarComponent, {
+            set: { changeDetection: ChangeDetectionStrategy.Default }
+         })
+         .overrideComponent(StSidebarItemListComponent, {
             set: { changeDetection: ChangeDetectionStrategy.Default }
          })
          .compileComponents();  // compile template and css
@@ -39,7 +43,14 @@ describe('StSidebar', () => {
          { id: 'identities', label: 'Identities' },
          { id: 'masters', label: 'Masters' },
          { id: 'agents', label: 'Agents' },
-         { id: 'roles', label: 'Roles' }
+         { id: 'roles', label: 'Roles' },
+         {
+            id: 'complex', label: 'Complex', items: [
+            { id: 'child-1', label: 'Child 1' },
+            { id: 'child-2', label: 'Child 2' },
+            { id: 'child-3', label: 'Child 3' }
+         ]
+         }
       ];
 
       fixture.detectChanges();
@@ -53,51 +64,106 @@ describe('StSidebar', () => {
    });
 
    it('by default, first item is active if the active item is not introduced', () => {
-      expect(itemList[0].classList).toContain('item__active');
+      expect(itemList[0].classList).toContain('item--active');
    });
 
    it('if active item is updated from outside, it has to be updated as active', () => {
       // by default, first item is active
-      expect(itemList[0].classList).toContain('item__active');
+      expect(itemList[0].classList).toContain('item--active');
 
       // active is changed from outside
       component.active = component.items[4].id;
       fixture.detectChanges();
 
-      expect(itemList[4].classList).toContain('item__active');
-      expect(itemList[0].classList).not.toContain('item__active');
+      expect(itemList[4].classList).toContain('item--active');
+      expect(itemList[0].classList).not.toContain('item--active');
 
    });
 
    describe('When user clicks on a tab', () => {
 
-      it('When user clicks on a tab, event is emitted if it was not already activated', () => {
-         spyOn(component.change, 'emit');
-         itemList[2].click();
+      describe('and it does not have children', () => {
 
-         expect(component.change.emit).toHaveBeenCalledWith(component.items[2].id);
+         it('event is emitted if it was not already activated', () => {
+            spyOn(component.change, 'emit');
+            (<HTMLSpanElement> itemList[2].querySelector('.item__label')).click();
 
-         (<jasmine.Spy> component.change.emit).calls.reset();
-         component.active = component.items[2].id;
+            expect(component.change.emit).toHaveBeenCalledWith(component.items[2].id);
 
-         itemList[2].click();
+            (<jasmine.Spy> component.change.emit).calls.reset();
+            component.active = component.items[2].id;
 
-         fixture.detectChanges();
-         expect(component.change.emit).not.toHaveBeenCalled();
+            (<HTMLSpanElement> itemList[2].querySelector('.item__label')).click();
 
-         itemList[3].click();
+            fixture.detectChanges();
+            expect(component.change.emit).not.toHaveBeenCalled();
 
-         fixture.detectChanges();
-         fixture.changeDetectorRef.markForCheck();
+            (<HTMLSpanElement> itemList[3].querySelector('.item__label')).click();
 
-         expect(component.change.emit).toHaveBeenCalledWith(component.items[3].id);
+            fixture.detectChanges();
+            fixture.changeDetectorRef.markForCheck();
+
+            expect(component.change.emit).toHaveBeenCalledWith(component.items[3].id);
+         });
+
+         it('this tab is displayed as active one', () => {
+            component.active = component.items[2].id;
+            fixture.detectChanges();
+
+            expect(itemList[2].classList).toContain('item--active');
+         });
       });
 
-      it('this tab is displayed as active one', () => {
-         component.active = component.items[2].id;
-         fixture.detectChanges();
+      describe('and it has children', () => {
+         it('if it is collapsed, these children are displayed', () => {
+            let itemWithChildren = itemList[5];
 
-         expect(itemList[2].classList).toContain('item__active');
+            expect(itemWithChildren.querySelectorAll('ul').length).toBe(0);
+
+            (<HTMLSpanElement> itemWithChildren.querySelector('.item__label')).click();
+
+            fixture.detectChanges();
+
+            expect(itemWithChildren.classList).not.toContain('item--active');
+            let childrenElements = itemWithChildren.querySelectorAll('li');
+
+            fixture.detectChanges();
+
+            expect(childrenElements.length).toBe(component.items[5].items.length);
+         });
+
+         it('but it is expanded, these children are hidden', () => {
+            let itemWithChildren = itemList[5];
+
+            expect(itemWithChildren.querySelectorAll('ul').length).toBe(0);
+
+            (<HTMLSpanElement> itemWithChildren.querySelector('.item__label')).click();
+
+            fixture.detectChanges();
+            expect(itemWithChildren.classList).toContain('item--expanded');
+
+            (<HTMLSpanElement> itemWithChildren.querySelector('.item__label')).click();
+
+            fixture.detectChanges();
+
+            let childrenElements = itemWithChildren.querySelectorAll('li');
+            expect(childrenElements.length).toBe(0);
+            expect(itemWithChildren.classList).not.toContain('item--expanded');
+         });
+
+         it('if user clicks on a child after clicking on its parent, it is displayed as active ' +
+            'and class "item-list--has-active" is added to its parent', () => {
+            const itemWithChildren = itemList[5];
+            (<HTMLSpanElement> itemWithChildren.querySelector('.item__label')).click();
+            fixture.detectChanges();
+            const childrenElements = itemWithChildren.querySelectorAll('li');
+
+            component.active = component.items[5].items[0].id; // active the first child
+            fixture.detectChanges();
+
+            expect(childrenElements[0].classList).toContain('item--active');
+            expect(itemWithChildren.classList).toContain('item--has-active');
+         });
       });
    });