+++ src/lib/st-sidebar/st-sidebar.component.ts
@@ -8,11 +8,11 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { ChangeDetectionStrategy, Component, Input, EventEmitter, Output, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, EventEmitter, Output, ChangeDetectorRef } from '@angular/core';
 import { StSidebarItem } from './st-sidebar-item.interface';
 
 /**
- * @description {Component} [Table]
+ * @description {Component} [Sidebar]
  *
  * The sidebar component has been designed to navigate through different sections of a web page.
  *
@@ -38,50 +38,20 @@ import { StSidebarItem } from './st-sidebar-item.interface';
    host: { class: 'st-sidebar' },
    changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class StSidebarComponent implements OnInit {
+export class StSidebarComponent {
    /** @Input {string} [title=''] Title displayed on the top of menu */
    @Input() title: string = '';
+   /** @Input {string} [active=''] The id of the current active item */
+   @Input() active: string = '';
    /** @Input {StSidebarItem[]} [items=''] List of items displayed on the menu */
    @Input() items: StSidebarItem[] = [];
    /** @Output {string} [change=''] Event emitted when the active item  is changed */
    @Output() change: EventEmitter<string> = new EventEmitter<string>();
 
-   private _active: string;
 
-   /** @Input {string} [active=''] The id of the current active item */
-   @Input() get active(): string {
-      return this._active;
-   }
-
-   ngOnInit(): void {
-      if (!this._active && this.items) {
-         this._active = this.items[0].id;
-      }
-   }
-
-   set active(active: string) {
-      if (active !== this._active) {
-         this._active = active;
-      }
-   }
-
-   getItemClasses(item: StSidebarItem): any {
-      let classes: any = {};
-      if (item.class) {
-         classes[item.class] = true;
-      }
-      classes.item__active = this.isActive(item.id);
-      return classes;
-   }
-
-   onSelectItem(itemId: string): void {
-      if (itemId !== this._active) {
+   onChange(itemId: string): void {
+      if (this.active !== itemId) {
          this.change.emit(itemId);
       }
    }
-
-   private isActive(itemId: string): boolean {
-      return this._active === itemId;
-   }
-
 }